Support monorepos correctly
I'm not sure if this is a hack, or the Right Way, but...

Consider a monorepo with a prefixed, multi-package Stack project:

    repo/
      something/
        .stack-work/
        stack.yaml
        package-a/
          .stack-work/
          package.yaml
          src/
        package-b/
          .stack-work/
          package.yaml
          src/

In the case that package-b depends on package-a, we need to read
package-a's mix files and source code when processing package-b's tix
files. To do so, as far as I can tell, requires changing into
package-a's directory during those steps. To know where to change to, as
far as I can tell, requires asking stack about local dependencies for
the package-id's we find in the tix files (e.g. package-a-0.0.0-abc123).

So, that's what we do.

In addition, this change:

- Removes --directory

  It turns out this is not needed (changing into package-b's directory)
  to support monorepos. All processing can occur from the top-level
  (something/), provided we switch to local dependency directories, and
  back, at the right times.

- Adds --prefix

  Processing must occur from the package root (something/) so the paths
  output will lack that prefix. Code Climate requires paths are relative
  to the project root (repo/), so we can support this by accepting and
  pre-pending "something/"

  NOTE: the trailing slack is required, so that the default value of ""
  doesn't result in a seemingly absolute path. I might work around this
  internally at some point.

- Removes handleIO (die . show)

  At some point, I thought this provided a nicer experience, but I'm
  starting to think it's no different than letting the exceptions throw
  all the way out.